+++ src-ui/src/lib/components/ui/setting/index.ts
@@ -5,4 +5,5 @@
 import { Toggle } from "svelte-5-ui-lib";
 import Setting from "./Setting.svelte";
 import SettingToggle from "./SettingToggle.svelte";
+
 export { Setting, SettingToggle, Toggle };