+++ src-ui/src/routes/settings/View.svelte
@@ -198,7 +198,7 @@ $effect(() => {
 
 <div class="space-y-3 w-full">
   <div>
-    <h3 class="text-lg font-medium">
+    <h3 class="font-semibold text-lg">
       {m.settingsViewDescription()}
     </h3>